# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Data Guardian API
  description: Base OpenAPI definition for Data Guardian
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Server
paths:
...
  /api/v1/users:
    get:
      tags:
        - User
      summary: Get all users
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
    post:
      tags:
        - User
      summary: Create a newly user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/v1/users/{id}/type:
    put:
      tags:
        - User
      summary: Assign user type to user
      operationId: assignUserType
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignUserTypeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Not Found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserType:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          description: Example "DONO_DE_RESTAURANTE" or "CLIENTE"
...
    UserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        userType:
          $ref: '#/components/schemas/UserType'
    AssignUserTypeRequest:
      type: object
      required:
        - userTypeId
      properties:
        userTypeId:
          type: integer
          format: int64